Remove Image/ImageCache from cyclegan.py and use factored classes.
